# gemma : allow offloading the output tensor