@cebtenzzre 2024-02-21
This change breaks CPU inference of the Q4_0 quant, with or without #5650:
```
llm_load_tensors: ggml ctx size =    0.10 MiB
ggml_new_object: not enough space in the context's memory pool (needed 101968, available 101600)
[1]    60495 segmentation fault (core dumped)
```