@slaren 2024-02-21
Should be fixed in https://github.com/ggerganov/llama.cpp/pull/5651